requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.5.0
python-dotenv==1.0.0
psycopg2-binary
drf-yasg==1.21.5
gunicorn==20.1.0
django-cors-headers==3.13.0
phonenumbers==8.13.8
django-phonenumber-field==7.0.2
